Create Read_mapping_with_BWA.sh
This one is dense. This script maps reads onto a draft version of the new reference genome. First I created a list of samples that I want to read. In this case, I was focusing on Iberian Littorina saxatilis. Then I change the job array iterator to be the sample name. This means that a seperate job will be run for each sample. I then copied the working data into a temporary directory to avoid memory issues. Once this is done I ran BWA-MEM for each sample with the default parameters. I have specified read group information with `-R`, but this is optional, it is only needed if you plan to use GATK. After BWA has finished I check the header to see if all scaffolds are present before sorting, indexing and summarising (flagstast) with samtools.