Create Variant_calling_with_DeepVariant.sh
This script was for a test run of variant calling with DeepVariant. Individual calls are run for each sample in a while loop. Note I am using a single contig of the old reference genome here. Individual calls were merged with `bcftools merge`, does not perform a joint calling step.